applications/mail/src/app/helpers/expiration.ts
import { getUnixTime } from 'date-fns';

import { MAILBOX_LABEL_IDS } from '@proton/shared/lib/constants';
import { UserModel } from '@proton/shared/lib/interfaces';
import { isFrozenExpiration } from '@proton/shared/lib/mail/messages';

import { MessageState } from '../logic/messages/messagesTypes';

export const canSetExpiration = (featureFlagValue: boolean, user: UserModel, messageState?: MessageState) => {
    const hasFrozenExpiration = isFrozenExpiration(messageState?.data);
    const { LabelIDs = [] } = messageState?.data || {};

    if (hasFrozenExpiration) {
        return false;
    }

    if (!featureFlagValue) {
        return false;
    }

    if (user.isFree) {
        return false;
    }

    if (!LabelIDs.length || LabelIDs.includes(MAILBOX_LABEL_IDS.SPAM) || LabelIDs.includes(MAILBOX_LABEL_IDS.TRASH)) {
        return false;
    }

    return true;
};

export const getExpirationTime = (date?: Date) => {
    return date ? getUnixTime(date) : null;
};
